requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
boto==2.49.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
conda==4.5.12
conda-build==3.17.6
conda-verify==3.1.1
contextlib2==0.5.5
cryptography==2.4.2
Django==2.1.5
django-braces==1.14.0
django-crispy-forms==1.8.1
docutils==0.14
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.6
jsonschema==2.6.0
keyring==17.0.0
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.26.0
msgpack==0.5.6
olefile==0.46
Pillow==5.3.0
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psutil==5.4.8
py==1.7.0
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.3.1
pylint==2.2.2
pyodbc==4.0.25
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.0.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
